+++ genSurround/shaders/MainPS.hlsl
@@ -78,7 +78,7 @@ float fnoisev3(float3 uv) {
     return result;
 }
 
-float lerp2(float fromA, float fromB, float toA, float toB, double value) {
+float lerp2(float fromA, float fromB, float toA, float toB, float value) {
     return ((value-fromA)/(fromB-fromA))*(toB-toA)+toA;
 }
